+++ src/score/engine/obj-bar-line.ts
@@ -1,7 +1,7 @@
 import { MusicObject } from "./music-object";
 import { View } from "./view";
 import { ObjMeasure } from "./obj-measure";
-import { MBarLineRight, MBarLineLeft, Navigation, MusicInterface, MStaffBarLine } from "../pub";
+import { MBarLineRight, MBarLineLeft, AnnotationKind, MusicInterface, MStaffBarLine } from "../pub";
 import { PlayerColumnProps } from "./player-engine";
 import { DocumentSettings } from "./settings";
 import { ObjNotationLine, ObjStaff } from "./obj-staff-and-tab";
@@ -213,10 +213,10 @@ export class ObjBarLineLeft extends ObjBarLine {
         let m = this.measure;
         let prev = m.getPrevMeasure();
 
-        if (m.hasNavigation(Navigation.StartRepeat)) {
+        if (m.hasAnnotationKind(AnnotationKind.StartRepeat)) {
             // If prev measure on same row has end-repeat:
             //     prev measure draws end-start-repeat and this measure does not draw start-repeat
-            if (prev && prev.row === m.row && prev.hasNavigation(Navigation.EndRepeat)) {
+            if (prev && prev.row === m.row && prev.hasAnnotationKind(AnnotationKind.EndRepeat)) {
                 return BarLineType.None;
             }
             else {
@@ -254,10 +254,10 @@ export class ObjBarLineRight extends ObjBarLine {
         let m = this.measure;
         let next = m.getNextMeasure();
 
-        if (m.hasNavigation(Navigation.EndRepeat)) {
+        if (m.hasAnnotationKind(AnnotationKind.EndRepeat)) {
             // If next measure on same row has start-repeat:
             //     this measure draws end-start-repeat and next measure does not draw start-repeat
-            if (next && next.row === m.row && next.hasNavigation(Navigation.StartRepeat)) {
+            if (next && next.row === m.row && next.hasAnnotationKind(AnnotationKind.StartRepeat)) {
                 return BarLineType.EndStartRepeat;
             }
             else {
@@ -272,12 +272,12 @@ export class ObjBarLineRight extends ObjBarLine {
             return BarLineType.Double;
         }
 
-        if (m === m.row.getLastMeasure() && next && next.row === m.row.getNextRow() && next.hasNavigation(Navigation.StartRepeat)) {
+        if (m === m.row.getLastMeasure() && next && next.row === m.row.getNextRow() && next.hasAnnotationKind(AnnotationKind.StartRepeat)) {
             // IF this is last measure of row && next row begins with start repeat
             return BarLineType.Double;
         }
 
-        if (next && next.hasNavigation(Navigation.StartRepeat)) {
+        if (next && next.hasAnnotationKind(AnnotationKind.StartRepeat)) {
             return BarLineType.None;
         }