+++ workspaces/features-demo/src/demo-pieces/staff-guitar.ts
@@ -17,7 +17,7 @@ export function createStaffConfigGuitarDemo() {
         .addNote(1, "C4", "8n", { string: 2 })
 
         .addMeasure()
-        .addNavigation("startRepeat")
+        .addAnnotation("startRepeat")
         .addChord(0, ["E4", "C3"], "8n", { string: [1, 5] }).addAnnotation("C", "chordLabel")
         .addNote(0, "E3", "8n", { string: 4 })
         .addNote(0, "G3", "8n", { string: 3 })